support sparsecootensortype
