+++ paddle/fluid/pir/dialect/operator/ir/op_type.cc
@@ -78,8 +78,55 @@ DenseTensorArrayType DenseTensorArrayType::dyn_cast_impl(Type type) {
   return nullptr;
 }
 
+pir::Type SparseCooTensorType::dtype() const { return storage()->dtype_; }
+
+const SparseCooTensorType::Dim& SparseCooTensorType::dims() const {
+  return storage()->dims_;
+}
+
+const SparseCooTensorType::Dim& SparseCooTensorType::meta_dims() const {
+  return storage()->meta_dims_;
+}
+
+DataLayout SparseCooTensorType::data_layout() const {
+  return storage()->layout_;
+}
+
+pir::DenseTensorType SparseCooTensorType::non_zero_indices() const {
+  return storage()->non_zero_indices_;
+}
+
+pir::DenseTensorType SparseCooTensorType::non_zero_elements() const {
+  return storage()->non_zero_elements_;
+}
+
+bool SparseCooTensorType::coalesced() const { return storage()->coalesced_; }
+
+bool SparseCooTensorType::classof(Type type) {
+  if (type) {
+    if (type.type_id() == type_id()) return true;
+    if (auto wrap_type = type.dyn_cast<pir::WrapTypeInterface>()) {
+      return classof(wrap_type.prim_type());
+    }
+  }
+  return false;
+}
+
+SparseCooTensorType SparseCooTensorType::dyn_cast_impl(Type type) {
+  if (type) {
+    if (type.type_id() == type_id()) {
+      return SparseCooTensorType(type.storage());
+    }
+    if (auto wrap_type = type.dyn_cast<pir::WrapTypeInterface>()) {
+      return dyn_cast_impl(wrap_type.prim_type());
+    }
+  }
+  return nullptr;
+}
+
 }  // namespace dialect
 }  // namespace paddle
 
 IR_DEFINE_EXPLICIT_TYPE_ID(paddle::dialect::SelectedRowsType)
 IR_DEFINE_EXPLICIT_TYPE_ID(paddle::dialect::DenseTensorArrayType)
+IR_DEFINE_EXPLICIT_TYPE_ID(paddle::dialect::SparseCooTensorType)
